Don't load everything up front
Return to a scheme much more like regular Yarr, though IMO with slightly
better terminology.  The list of articles displayed is called
a "snapshot" and the view preferences are called "snapshot parameters"
(which are taken from the query string).  As a temporary hack, any time
the parameters change the page is reloaded with an updated querystring.

The first 10 articles in the snapshot are included with the initial page
load.  Dynamically loading the rest is the only big piece of missing
functionality at this point.  Eventually they should also be lazily
added to the DOM to prevent it from getting too huge.