Less flashing on load
Fix some confusion where a loading snapshot caused display of a "404"
message for the article. The reduction of flashing is done by
altering the article view to display the top and bottom bars while the
snapshot loads, which unfortunately involved making the read/unread and
fave/unfave buttons understand null as a third state (which produces
a disabled and invisible button). It seems rather messy to push all of
this down to the buttons, but otherwise the focus may or may not shift
around when navigating next/prev between articles, depending on whether
the article has already been loaded. This way, the focus never moves if
within the top or bottom bar.