Remove swipe support
While it seemed to work on a Linux machine with a Wacom tablet, this
doesn't actually work in Firefox on a Windows 10 tablet with
a multi-touch screen. The event stream is cancelled almost immediately
as the browser interprets it as a potential scroll gesture. While it
might be possible to deal with this via liberal application of
JavaScript, I think that it is more important for scrolling to remain
performant than to support this swipe gesture.