Set ATOMIC_REQUESTS = True
The test environment in django.test.TestCase is within a transaction, so
just to be safe we'll use transactions in our views rather than the
default of autocommit.

This is prompted by a bug in the triggers I observed once but cannot
reproduce: a bulk "mark read" operation caused the unread counter to go
negative.