Track Feed.last_changed
This replaces Feed.last_updated, which was never properly tracked. Since
it's not clear what "updated" means (Does that mean adding an article?
Changing the feed title? What about updating an article?) this is simply
the timestamp that the feed's bytes last changed (i.e., there was
a MaybeUpdated poll result). While this produces a nonsensical result
when the feed includes a dynamic timestamp, such feeds are rare.