Generate an internally-compressed .ico
Rather than using icotools to generate an uncompressed .ico file (which
contains ~BMP images internally), generate one with embedded PNGs. This
is a few hundred bytes larger than the result of applying Brotli
over-the-top of the uncompressed .ico, but will enable embedding a larger
icon.

As a side effect, icotools (last release: 2018-03-07) is no longer
required.